[FEATURE] Make the ext friendly to EXT:container
EXT:container uses a pre-processing hook in DataHandler to do adjustments to data. It checks for the 'localize' command, so we cannot use our custom 'deepl' command. Thus use the post-processing hook and request parameters to make sure that container does all it has to do and process records after actual localization by TYPO3. This change also removes internal knowledge of the DataHandler processing and makes it less dependent on DataHandler internals.